+++ src/Components/ButtonOpenModal/ButtonOpenModal.module.scss
@@ -39,14 +39,22 @@
   color: $text-color;
   margin-right: 7px;
   cursor: default;
-
+  font-size: 14px;
+  line-height: 1;
+  @media screen and (min-width: 768px) {
+    font-size: 11px;
+  }
   &.visible {
     cursor: pointer;
   }
 
   &.difficulty {
+    font-size: 16px;
     text-transform: lowercase;
     color: $secondary-text;
+    @media screen and (min-width: 768px) {
+      font-size: 14px;
+    }
     &::first-letter {
       text-transform: uppercase;
     }
